requirements.txt
transformers 
diffusers
Pillow
gradio 